refactor: drop deprecated historical implementations
