+++ lib/bindings/ffmpeg.zod.ts
@@ -2,7 +2,7 @@ import {readlines} from '../util.ts'
 import { InputError, CommandError } from '../errors.ts'
 import { parse_duration } from '../parsers/duration.zod.ts'
 import type { Context } from '../context.ts'
-import type { Timestamp } from '../template_input.ts'
+import type { Timestamp } from '../template_input.zod.ts'
 
 type OnReadLine = (line: string) => void
 async function exec(cmd: string[]) {
