+++ lib/canvas/font.zod.ts
@@ -5,7 +5,6 @@ import * as culori from 'culori'
 import { compute_size } from '../geometry.zod.ts'
 import type { Context } from '../context.ts'
 import type { TextClipParsed, MediaClipParsed } from '../parsers/template.zod.ts'
-import { ContextExtended } from './round-rect.ts'
 
 // first in the list gets highest priority
 const PRIORITIED_FONT_FAMILIES = [
