+++ lib/zoompan.zod.ts
@@ -1,9 +1,8 @@
 import { InputError } from './errors.ts'
 import { parse_unit, parse_percentage } from './parsers/unit.ts'
 import { parse_duration } from './parsers/duration.zod.ts'
-import { AbstractClipMap } from './parsers/template.ts'
-import type { ClipID } from './template_input.ts'
-import type * as template_parsed from './parsers/template.ts'
+import { AbstractClipMap } from './util.ts'
+import type { ClipID } from './template_input.zod.ts'
 import type { ClipInfoMap } from './probe.zod.ts'
 import type { ClipGeometryMap } from './geometry.zod.ts'
 import type {Context} from './context.ts'
